invoices/requirements.txt
pymongo==4.6.1
minio==7.2.3
python-dotenv==1.0.1
Jinja2==3.1.3
weasyprint==52.5 